Split xtask workflow responsibilities
Repository-local xtask crates should own task semantics and workflow dispatch, while shared crates expose only WRAC build operations and generic workflow primitives.
