+++ Cargo.toml
@@ -7,7 +7,8 @@ members = [
     "crates/wrac_host_context",
     "crates/wrac_log",
     "crates/wrac_manifest",
-    "crates/wrac_xtask",
+    "crates/xtask_workflow",
+    "crates/wrac_build_ops",
     "crates/wrac_wxp_gui",
 ]
 resolver = "2"
@@ -17,8 +18,10 @@ clap-sys = "0.5.0"
 novonotes_run_loop = { git = "https://github.com/novonotes/wxp.git", package = "novonotes_run_loop", rev = "93d7ed412b7206351100e6a1970628d3bedab2cc" }
 run_loop_timer = { git = "https://github.com/novonotes/wxp.git", package = "run_loop_timer", rev = "93d7ed412b7206351100e6a1970628d3bedab2cc" }
 wrac_clap_adapter = { path = "crates/wrac_clap_adapter" }
+wrac_build_ops = { path = "crates/wrac_build_ops" }
 wrac_manifest = { path = "crates/wrac_manifest" }
 wrac_host_context = { path = "crates/wrac_host_context" }
 wrac_log = { path = "crates/wrac_log" }
 wrac_wxp_gui = { path = "crates/wrac_wxp_gui" }
+xtask_workflow = { path = "crates/xtask_workflow" }
 wxp = { git = "https://github.com/novonotes/wxp.git", rev = "93d7ed412b7206351100e6a1970628d3bedab2cc" }
