+++ crates/wrac_build_ops/src/validation/checks.rs
