+++ crates/wrac_build_ops/src/validation/report.rs
